factor out self-hosted benchmark and track_ functions for asv
